requirements.txt
crewai
crewai-tools
langchain-openai
pandas
numpy
matplotlib
seaborn
scikit-learn
requests
streamlit
python-dotenv  # <<< ESSENCIAL PARA CARREGAR O .ENV